Inoptic

Documentation for Inoptic.

Inoptic.Inoptic

Inoptic.InflationOptimalABSME2021

Inoptic.InflationOptimalABSME2022

Inoptic.InflationOptimalCORR2021

Inoptic.InflationOptimalCORR2022

Inoptic.InflationOptimalFixedExclusionABSME2026

Inoptic.InflationOptimalFixedExclusionCORR2024B

Inoptic.InflationOptimalFixedExclusionCORR2026

Inoptic.InflationOptimalFixedExclusionRMSE2026

Inoptic.InflationOptimalMSE2021

Inoptic.InflationOptimalMSE2022

Inoptic.InflationOptimalMSE2023

Inoptic.InflationOptimalMSE2024

Inoptic.InflationOptimalMSE2024B

Inoptic.InflationOptimalRMSE2026

Inoptic.InflationOptimalRMSE2026B

Inflation Optimal Combination Functions

The Inoptic package provides functions to compute optimal inflation measures based on the methodologies developed by the DIE-BG research group.

Inoptic.Inoptic — Module

Inoptic.jl

A Julia package to save, load the optimal inflation measures based on the methodologies developed by the DIE-BG research group.

Inoptic.InflationOptimalABSME2021 — Function

InflationOptimalABSME2021 <: CombinationFunction

Core inflation measure, optimal linear combination based on the ABSME methodology for the 2021 edition.

Inoptic.InflationOptimalABSME2022 — Function

InflationOptimalABSME2022 <: CombinationFunction

Core inflation measure, optimal linear combination based on the ABSME methodology for the 2022 edition.

Inoptic.InflationOptimalCORR2021 — Function

InflationOptimalCORR2021 <: CombinationFunction

Core inflation measure, optimal linear combination based on the CORR methodology for the 2021 edition.

Inoptic.InflationOptimalCORR2022 — Function

InflationOptimalCORR2022 <: CombinationFunction

Core inflation measure, optimal linear combination based on the CORR methodology for the 2022 edition.

Inoptic.InflationOptimalFixedExclusionABSME2026 — Function

InflationOptimalFixedExclusionABSME2026 <: InflationSpliceUnweighted

Inoptic.InflationOptimalFixedExclusionCORR2024B — Function

InflationOptimalFixedExclusionCORR2024B <: InflationSpliceUnweighted

Inoptic.InflationOptimalFixedExclusionCORR2026 — Function

InflationOptimalFixedExclusionCORR2026 <: InflationSpliceUnweighted

Inoptic.InflationOptimalFixedExclusionRMSE2026 — Function

InflationOptimalFixedExclusionRMSE2026 <: InflationSpliceUnweighted

Inoptic.InflationOptimalMSE2021 — Function

InflationOptimalMSE2021 <: CombinationFunction

Core inflation measure, optimal linear combination based on the MSE methodology for the 2021 edition.

Inoptic.InflationOptimalMSE2022 — Function

InflationOptimalMSE2022 <: CombinationFunction

Core inflation measure, optimal linear combination based on the MSE methodology for the 2022 edition.

Inoptic.InflationOptimalMSE2023 — Function

InflationOptimalMSE2023 <: CombinationFunction

Core inflation measure, optimal linear combination based on the MSE methodology for the 2023 edition.

Inoptic.InflationOptimalMSE2024 — Function

inflation_optmse2024 <: CombinationFunction Optimate MSE 2024 combination inflation function.

Inoptic.InflationOptimalMSE2024B — Function

InflationOptimalMSE2024B <: InflationSpliceUnweighted

Core inflation measure, optimal linear combination based on the MSE methodology for the 2024B edition.

Inoptic.InflationOptimalRMSE2026 — Function

InflationOptimalRMSE2026 <: InflationSpliceUnweighted

Core inflation measure, optimal linear combination based on the RMSE methodology for the 2026 edition.

Inoptic.InflationOptimalRMSE2026B — Function

InflationOptimalRMSE2026B <: InflationSpliceUnweighted

Core inflation measure, optimal linear combination based on the RMSE methodology for the 2026 edition. We use a complexity-based penalty for each kind of measure to assign more weight to the "robust" (simpler) measures